# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Твой API Справочник Организаций",
        "description": "API для работы с организациями, зданиями и видами деятельности",
        "contact": {
            "email": "твой@email.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
        "/api/buildings/organizations": {
            "get": {
                "tags": [
                    "Здания"
                ],
                "summary": "Получение списка организаций в здании",
                "description": "Возвращает данные здания, включая связанные организации и их телефонные номера.",
                "operationId": "getBuildingOrganizations",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "building_id",
                        "in": "query",
                        "description": "Идентификатор здания.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Данные здания с организациями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BuildingWithOrganizationsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Здание не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The building id field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "building_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The building id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/by-id": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получение информации об организации по идентификатору",
                "description": "Возвращает детальные данные организации, включая телефоны, здание и дерево видов деятельности.",
                "operationId": "getOrganizationById",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор организации.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 7
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Детальные данные организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FullOrganizationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Организация не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The id field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/activities/organizations": {
            "get": {
                "tags": [
                    "Виды деятельности"
                ],
                "summary": "Получение списка организаций по виду деятельности",
                "description": "Находит вид деятельности по названию и возвращает связанные организации с телефонными номерами. Сначала выполняется точный поиск, при отсутствии результатов — частичный поиск с сортировкой по длине названия.",
                "operationId": "getOrganizationsByActivity",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "activity_name",
                        "in": "query",
                        "description": "Название вида деятельности (минимум 3 символа).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        },
                        "example": "Рестораны"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Данные вида деятельности с организациями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityWithOrganizationsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Вид деятельности не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The activity name must be at least 3 characters."
                                        },
                                        "errors": {
                                            "properties": {
                                                "activity_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The activity name must be at least 3 characters."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ActivityResource": {
                "title": "Ресурс вида деятельности",
                "description": "Структура данных, представляющая вид экономической деятельности с иерархической организацией",
                "required": [
                    "id",
                    "name",
                    "depth"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор",
                        "description": "Уникальный числовой идентификатор вида деятельности",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "Наименование",
                        "description": "Официальное название вида экономической деятельности",
                        "type": "string",
                        "example": "Производство промышленного оборудования"
                    },
                    "depth": {
                        "title": "Уровень вложенности",
                        "description": "Порядковый уровень в иерархической структуре видов деятельности (0 - корневой уровень)",
                        "type": "integer",
                        "example": 0
                    },
                    "children": {
                        "title": "Дочерние элементы",
                        "description": "Массив подчиненных видов деятельности в иерархической структуре (рекурсивная структура)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityResource"
                        }
                    }
                },
                "type": "object"
            },
            "ActivityWithOrganizationsResource": {
                "title": "Ресурс вида деятельности с организациями",
                "description": "Представление вида экономической деятельности с привязанными организациями",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ActivityResource"
                    },
                    {
                        "properties": {
                            "activity": {
                                "title": "Данные вида деятельности",
                                "description": "Структура, содержащая основные параметры вида деятельности и связанные организации",
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/ActivityResource"
                                    },
                                    {
                                        "properties": {
                                            "organizations": {
                                                "title": "Связанные организации",
                                                "description": "Перечень организаций, осуществляющих данный вид экономической деятельности",
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/components/schemas/OrganizationWithPhonesResource"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "BuildingResource": {
                "title": "Ресурс здания",
                "description": "Структура данных, представляющая информацию о здании",
                "required": [
                    "id",
                    "address",
                    "coordinates"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор здания",
                        "description": "Уникальный числовой идентификатор объекта недвижимости",
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "title": "Адрес",
                        "description": "Физический адрес местонахождения здания",
                        "type": "string",
                        "example": "ул. Тверская, 15"
                    },
                    "coordinates": {
                        "title": "Географические координаты",
                        "description": "Геопозиция здания в системе координат WGS 84",
                        "required": [
                            "lat",
                            "lng"
                        ],
                        "properties": {
                            "lat": {
                                "title": "Широта",
                                "description": "Географическая широта в десятичных градусах",
                                "type": "number",
                                "format": "float",
                                "example": 55.7558
                            },
                            "lng": {
                                "title": "Долгота",
                                "description": "Географическая долгота в десятичных градусах",
                                "type": "number",
                                "format": "float",
                                "example": 37.6173
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "BuildingWithOrganizationsResource": {
                "title": "Ресурс здания с организациями",
                "description": "Структура данных, представляющая информацию о здании с привязанными организациями",
                "properties": {
                    "building": {
                        "title": "Данные здания",
                        "description": "Объект, содержащий информацию о здании и связанных с ним организациях",
                        "type": "object",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/BuildingResource"
                            },
                            {
                                "properties": {
                                    "organizations": {
                                        "title": "Организации в здании",
                                        "description": "Список организаций, размещенных в данном здании",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/OrganizationWithPhonesResource"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "FullOrganizationResource": {
                "title": "Полный ресурс организации",
                "description": "Комплексное представление данных об организации",
                "properties": {
                    "organization": {
                        "title": "Данные организации",
                        "description": "Объект, содержащий полную информацию об организации",
                        "type": "object",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/OrganizationResource"
                            },
                            {
                                "required": [
                                    "phones",
                                    "building",
                                    "activities"
                                ],
                                "properties": {
                                    "phones": {
                                        "title": "Контактные телефоны",
                                        "description": "Список контактных номеров телефонов организации",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/OrganizationPhoneResource"
                                        }
                                    },
                                    "building": {
                                        "$ref": "#/components/schemas/BuildingResource"
                                    },
                                    "activities": {
                                        "title": "Виды деятельности",
                                        "description": "Перечень видов экономической деятельности организации",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/ActivityResource"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "OrganizationPhoneResource": {
                "title": "Ресурс телефонного номера организации",
                "description": "Представление контактного телефонного номера организации",
                "required": [
                    "id",
                    "number"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор",
                        "description": "Уникальный числовой идентификатор телефонного номера",
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "title": "Номер телефона",
                        "description": "Контактный телефонный номер в международном или национальном формате",
                        "type": "string",
                        "example": "+7 (495) 123-45-67"
                    }
                },
                "type": "object"
            },
            "OrganizationResource": {
                "title": "Базовый ресурс организации",
                "description": "Минимальный набор данных для идентификации организации",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор организации",
                        "description": "Уникальный числовой идентификатор организации в системе",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "Наименование",
                        "description": "Официальное название организации",
                        "type": "string",
                        "example": "ООО 'Промышленные технологии'"
                    }
                },
                "type": "object"
            },
            "OrganizationWithPhonesResource": {
                "title": "Ресурс организации с телефонами",
                "description": "Данные организации, включая идентификатор, наименование и список телефонных номеров.",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/OrganizationResource"
                    },
                    {
                        "properties": {
                            "phones": {
                                "title": "Телефоны",
                                "description": "Список телефонных номеров, связанных с организацией.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrganizationPhoneResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "query"
            }
        }
    },
...
}